Refactor COCORecall
COCOBase is not feasible due to the way COCOMaP
must be computed.  The good news that comes
with this is that I am able to drastically
simplify the COCORecall implementation, as well
as make it more specialized.

This PR also updates the way constants on the metric
are tracked in order to support the default `merge_state`
method, provides more thorough documentation, and
merges COCOBase and COCORecall's unit tests
