+++ keras_cv/metrics/coco/utils.py
@@ -96,3 +96,76 @@ def sort_bboxes(boxes, axis=5):
         )
 
     return boxes_sorted_list.stack()
+
+
+def match_boxes(y_true, y_pred, threshold, ious):
+    """matches bounding boxes from y_true to boxes in y_pred.
+
+    Args:
+        y_true: bounding box tensor of shape [num_boxes, 4+].
+        y_pred: bounding box tensor of shape [num_boxes, 4+].
+        threshold: minimum IoU for a pair to be considered a match.
+        ious: lookup table from [y_true, y_pred] => IoU.
+    Returns:
+        indices of matches between y_pred and y_true.
+    """
+    num_true = tf.shape(y_true)[0]
+    num_pred = tf.shape(y_pred)[0]
+
+    gt_matches = tf.TensorArray(
+        tf.int32,
+        size=num_true,
+        dynamic_size=False,
+        infer_shape=False,
+        element_shape=(),
+    )
+    pred_matches = tf.TensorArray(
+        tf.int32,
+        size=num_pred,
+        dynamic_size=False,
+        infer_shape=False,
+        element_shape=(),
+    )
+    for i in tf.range(num_true):
+        gt_matches = gt_matches.write(i, -1)
+    for i in tf.range(num_pred):
+        pred_matches = pred_matches.write(i, -1)
+
+    for detection_idx in tf.range(num_pred):
+        match_index = -1
+        iou = tf.math.minimum(threshold, 1 - 1e-10)
+
+        for gt_idx in tf.range(num_true):
+            if gt_matches.gather([gt_idx]) > -1:
+                continue
+            # TODO(lukewood): update clause to account for gtIg
+            # if m > -1 and gtIg[m] == 0 and gtIg[gind] == 1:
+
+            if not ious[gt_idx, detection_idx] >= iou:
+                continue
+
+            iou = ious[gt_idx, detection_idx]
+            match_index = gt_idx
+
+        # Write back the match indices
+        pred_matches = pred_matches.write(detection_idx, match_index)
+        if match_index == -1:
+            continue
+        gt_matches = gt_matches.write(match_index, detection_idx)
+    return pred_matches.stack()
+
+def result(self):
+    present_values = self.ground_truth_boxes != 0
+    n_present_categories = tf.math.reduce_sum(
+        tf.cast(present_values, tf.float32), axis=-1
+    )
+    if n_present_categories == 0.0:
+        return 0.0
+
+    recalls = tf.math.divide_no_nan(
+        self.true_positives, self.ground_truth_boxes[None, :]
+    )
+    recalls_per_threshold = (
+        tf.math.reduce_sum(recalls, axis=-1) / n_present_categories
+    )
+    return tf.math.reduce_mean(recalls_per_threshold)
